Fix possible overlapping UV coordinates
CURA-12528
The initial mesh data contains individual vertices coordinates for each face, and this is required for the charts packing to work properly. However there are some charts that contain non-linked faces, so we temporarily create a list of unique vertices indices, and use it to split the charts groups in sub-groups with only linked faces.